nav_twin/requirements.txt
# Nav Twin - Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ML/AI
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Transit Data
gtfs-realtime-bindings==1.0.0
requests==2.31.0

# Google Maps
googlemaps==4.10.0

# OpenAI (for NLP)
openai==1.3.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
httpx==0.25.1

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Security
cryptography==41.0.7
